Refactor : 컴포넌트 분리, 커스텀 훅 생성, 구조 리팩토링
