+++ src/components/buttons/CustomDateButton.jsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export default function CustomDateButton() {
+  return <Wrapper>달력</Wrapper>;
+}
+
+const Wrapper = styled.div`
+  grid-area: datebutton;
+`;
