+++ src/components/buttons/ResetButton.jsx
@@ -1,28 +1,32 @@
 import styled from "styled-components";
 
-export default function ResetButton({ setUploadedInfo }) {
+export default function ResetButton({ uploadedInfo, setUploadedInfo }) {
   return (
-    <Wrapper>
-      <button className="reset" onClick={() => setUploadedInfo(null)}>
-        RESET
-      </button>
-    </Wrapper>
+    uploadedInfo?.lastModified && (
+      <Wrapper>
+        <button className="reset" onClick={() => setUploadedInfo(null)}>
+          RESET
+        </button>
+      </Wrapper>
+    )
   );
 }
 
 const Wrapper = styled.div`
+  grid-area: reset;
+  display: flex;
+  justify-content: center;
+
   .reset {
     font-family: "DOSGothic";
     background: #bd4030;
     box-shadow: 0px 7px 0px 0px #962b1f;
 
-    display: block;
     position: relative;
     float: left;
     width: 60px;
     height: 50px;
-    padding: 0;
-    margin: 10px 20px 10px 0;
+    margin: 10px 0 10px 0;
     text-align: center;
     font-size: 18px;
